Add alert text for non-PWD's
A button has been added for PWD's, but some users play and click this
button that basically will call the forPWD() and will eventually lead
to more orders in the sqlite database instances that will lead to a
big overhead.

PS: just practicing being a PRO in git commit messages.